Fix/51

+++ libs/scooter-tagore/src/lib/Component.jsx
@@ -1,5 +1,5 @@
 import { NodeViewWrapper } from "@tiptap/react";
-import React from "react";
+import React, { useEffect } from "react";
 import axios from "axios";
 import { Input } from "@factly/scooter-ui";
 import TagoreCommandsMenu from "./Menu";
@@ -18,6 +18,7 @@ export const TagoreComponent = props => {
     editor,
     node,
     menuIndex = 0,
+    deleteNode,
     extension: {
       options: {
         // apiUrl = "http://localhost:8080",
@@ -25,6 +26,8 @@ export const TagoreComponent = props => {
         menuItems = {},
         //systemPrompt=` Return the response as a valid HTML without html, head or body tags`,
         fetcher,
+        sse,
+        stream,
       } = {},
     },
   } = props;
@@ -82,7 +85,9 @@ export const TagoreComponent = props => {
       setError(null);
       setContent("");
       const data = await fetcher(
-        content ? `${input} ${content}` : input || inputValue.split(":")[1],
+        content
+          ? `${input} ${content}`
+          : input || inputValue.split(":")[1] || inputValue,
         selectedOption
       );
 
@@ -96,13 +101,60 @@ export const TagoreComponent = props => {
       setLoading(false);
       setGenerated(true);
       setInputValue("");
-      return data;
+      const modifiedData = {
+        ...data,
+        output: data.output.replace(/\n|\t|(?<=>)\s*/g, ""),
+      };
+      return modifiedData;
     } catch (error) {
       hideMenu();
       setLoading(false);
       setError(error);
       console.log("Couldn't fetch data");
     }
+  }
+
+  async function streamData(input, selectedOption, callback) {
+    setLoading(true);
+    setGenerated(false);
+    setError(null);
+    setContent("");
+    let source = sse(input, selectedOption);
+
+    source.addEventListener("message", event => {
+      let text = JSON.parse(event.data);
+      setContent(text.output);
+    });
+
+    source.addEventListener("error", event => {
+      source.close();
+      setLoading(false);
+      callback(); //showmenu
+
+      if (!String(event.data).includes("[DONE]")) {
+        setError(true);
+        hideMenu();
+
+        setLoading(false);
+        console.log("Couldn't fetch data");
+        return;
+      } else {
+        setGenerated(true);
+      }
+    });
+
+    source.stream();
+
+    // setLoading(false);
+
+    setInputValue("");
+
+    // catch (error) {
+    //   hideMenu();
+    //   setLoading(false);
+    //   setError(error);
+    //   console.log("Couldn't fetch data");
+    // }
   }
 
   function filterByTitle(arr, title = "", showImprovingOptions = false) {
@@ -225,19 +277,21 @@ export const TagoreComponent = props => {
 
   const handleEnter = async e => {
     if (e.key === "Enter") {
-      const data = await fetchData(
-        inputValue.split(":")[1],
-        currentSelectedItem.promptId
-      );
-      editor.commands.insertContentAt(
-        props.getPos(),
-        //props.getPos(),
-        `${data.output}`
-      );
+      // const data = await fetchData(
+      //   inputValue.split(":")[1] || inputValue,
+      //   currentSelectedItem?.promptId || "default"
+      // );
+      // console.log("data @@@@@@@@@@", data);
+      // editor.commands.insertContentAt(
+      //   props.getPos(),
+      //   //props.getPos(),
+      //   `${data.output.replace(/\n|\t|(?<=>)\s*/g, "")}`
+      // );
 
-      hideMenu();
-      props.deleteNode();
-      setContent("");
+      // hideMenu();
+      // props.deleteNode();
+      // setContent("");
+      handleSubmit();
     }
     if (e.key === "Escape") {
       hideMenu();
@@ -252,20 +306,63 @@ export const TagoreComponent = props => {
     }
   };
 
-  const handleSubmit = async () => {
-    const data = await fetchData(
-      inputValue.split(":")[1],
-      currentSelectedItem.promptId
-    );
-    editor.commands.insertContentAt(
-      props.getPos(),
-      //props.getPos(),
-      `${data.output}`
-    );
+  function selectElementContents(el) {
+    if (window.getSelection && document.createRange) {
+      const sel = window.getSelection();
+      const range = document.createRange();
+      range.selectNodeContents(el);
+      sel.removeAllRanges();
+      sel.addRange(range);
+    } else if (document.selection && document.body.createTextRange) {
+      const textRange = document.body.createTextRange();
+      textRange.moveToElementText(el);
+      textRange.select();
+    }
+  }
 
-    hideMenu();
-    props.deleteNode();
-    setContent("");
+  const handleSubmit = async () => {
+    if (stream) {
+      setContent("");
+      setLoading(true);
+      setError(false);
+
+      let source = sse(
+        inputValue.split(":")[1] || inputValue,
+        currentSelectedItem?.promptId || "default"
+      );
+
+      source.addEventListener("message", event => {
+        let text = JSON.parse(event.data);
+        setContent(text.output);
+      });
+
+      source.addEventListener("error", event => {
+        source.close();
+        setLoading(false);
+        if (!String(event.data).includes("[DONE]")) {
+          setError(true);
+          return;
+        } else {
+          return;
+        }
+      });
+
+      source.stream();
+    } else {
+      const data = await fetchData(
+        inputValue.split(":")[1] || inputValue,
+        currentSelectedItem?.promptId || "default"
+      );
+      if (data) {
+        //    console.log({ "g": "generate", data })
+        setContent(data.output.replace(/\n|\t|(?<=>)\s*/g, ""));
+        showMenu();
+      }
+
+      hideMenu();
+      props.deleteNode();
+      setContent("");
+    }
   };
 
   const showMenu = () => {
@@ -293,6 +390,49 @@ export const TagoreComponent = props => {
 
   //getPosition();
 
+  const ref = React.useRef(null);
+  const [isOpen, setIsOpen] = React.useState(true);
+  React.useEffect(() => {
+    function handleClickOutside(event) {
+      if (ref.current && !ref.current.contains(event.target)) {
+        setIsOpen(false);
+      }
+    }
+
+    if (isOpen) {
+      document.addEventListener("mousedown", handleClickOutside);
+    } else {
+      if (generated) {
+        props.editor.commands.insertContentAt(
+          { from, to },
+          //props.getPos(),
+          content.replace(/\n|\t|(?<=>)\s*/g, "")
+        );
+        props.deleteNode();
+      } else {
+        props.deleteNode();
+      }
+      document.removeEventListener("mousedown", handleClickOutside);
+    }
+
+    return () => {
+      document.removeEventListener("mousedown", handleClickOutside);
+    };
+  }, [isOpen]);
+
+  useEffect(() => {
+    if (generated) {
+      const element = document.querySelector(".generated-content");
+      selectElementContents(element);
+    } else {
+      window.getSelection().removeAllRanges();
+    }
+
+    return () => {
+      // window.getSelection().removeAllRanges();
+    };
+  }, [generated]);
+
   return (
     <NodeViewWrapper
       className={
@@ -305,6 +445,7 @@ export const TagoreComponent = props => {
           // }px)`,
         }
       }
+      ref={ref}
     >
       {/* { console.log(`translate(${getPosition()?.left}px, ${getPosition()?.bottom}px)`)} */}
       <div className="content">
@@ -314,12 +455,14 @@ export const TagoreComponent = props => {
         ></div>
         {error && (
           <div
-            className="generated-content"
+            className="tagore-error-container"
             dangerouslySetInnerHTML={{ __html: "try again" }}
           ></div>
         )}
 
-        {loading && <div className="generated-content">Generating...</div>}
+        {loading && (
+          <div className="tagore-loading-container">Generating...</div>
+        )}
         <div className="tagore-input-container">
           <BsStars className="icon" size={"1rem"} />
           <input
@@ -330,7 +473,7 @@ export const TagoreComponent = props => {
             placeholder="Ask AI to write anything..."
             onFocus={showMenu}
             //onBlur={}
-            onSubmit={fetchData}
+            onSubmit={handleSubmit}
             ref={inputRef}
             autoFocus={true}
           />
@@ -359,6 +502,9 @@ export const TagoreComponent = props => {
           inputValue={inputValue}
           setInputValue={setInputValue}
           fetchData={fetchData}
+          streamData={streamData}
+          setGenerated={setGenerated}
+          stream={stream}
           setContent={setContent}
           selectedContent={selectedContent}
           to={to}
